Write relocation address into SymbolRelocs' memory
Previously, I had assumed that when a binary calls a `SymbolReloc` relocation
(e.g., calling a `GLOB_DAT` relocation directly when a binary is compiled with
`-fno-plt`), then it would be calling the address of the relocation. This is
not actually true, however: it calls the memory that the relocation address
points to. This distinction is critical, as we had previously been populating
`GLOB_DAT` relocations with entirely symbolic bytes—see the hack described in
`Note [Subtleties of resolving PLT stubs] (Wrinkle 1: .plt.got)` in
`Grease.Macaw.PLT`. This means that `grease` was unable to resolve external
function calls that go through `GLOB_DAT` relocations in this way.

This commit refines the hack such that instead of writing entirely symbolic
bytes to `SymbolReloc` relocations' memory, we instead write the address of the
relocation itself. This is still a hack, but it is a darn useful one. The
relocation address is a reasonably unique value that can be used to look up
when a function call corresponds to a `SymbolReloc` address, which allows us to
treat them as external function calls.

I have only observed `gcc` calling into `GLOB_DAT` relocations directly on
x86-64. On AArch32 and PPC, `gcc` compiles them into calls into `JUMP_SLOT`
relocations instead, which bypasses the issue entirely. Unfortunately, this
makes it not straightforward to add a test case, as this property critically
depends on a dynamically linked binary, which `grease`'s test suite does not
currently support for x86-64 binaries (see #30).

Fixes #62.