Move grease code to dedicated subdirectory
Instead of having the `grease` code live in separate, top-level `src`, `test`,
and `main` directories, this puts all of the `grease` code under a dedicated
`grease` subdirectory.

This is purely a refactoring in preparation for #46. In subsequent commits, we
will split up the code under the `grease` subdirectory into further
sub-libraries.