Split out grease-cli package
This splits out the core of the CLI bits of `grease` into a separate
`grease-cli` package. At a first approximation, the CLI-specific bits are the
`Grease.Main` module, the `Grease.Profiler.*` bits, the `grease` executable,
and the `grease-tests` test suite (which depends on `Grease.Main`). This leaves
the `grease` package as a library-only package.

This is progress towards a fix for #46. In a subsequent commit, we will split
the `grease` package into further, architecture-specific packages.