Fix a panic involving `fresh-bytes`
The message was:

    Location:  GlobalState.globalMuxFn
    Message:   Attempting to merge global states of incorrect branch depths:

I tracked this down to the call in
`Lang.Crucible.Simulator.ExecutionTree.abortedGlobals`. I think that
function applies `globalMuxFn` when its preconditions are not met. This
commit avoids muxing globals at all, because we're only interested in a
single branch for the sake of concretization anyway.

Unfortunately, I have not yet been able to come up with a standalone
test-case.