Add --overrides-yaml
This adds a way to specify function overrides for binaries by address (instead
of by name, as is the case with `--overrides`). The new `--overrides-yaml`
flag takes a YAML file that looks like this:

```yaml
function address overrides:
  ADDR_1: "NAME_1"
  ...
  ADDR_n: "NAME_n"
```

For more details, see the new documentation in `doc/overrides.md`.

Fixes #20.