ci: Reorder steps for faster failures
Solvers are a runtime dependency of the test suite, so they don't need
to be installed if e.g., `grease` doesn't type-check.

Our Cabal files rarely change, so `cabal check` almost always passes, so
it's safe to put it at the end to get faster feedback from the build and
test stages.