cli: Unify handling of enumerations
Previously, we had two ways of signaling possible values for
enumerations. One was to add a parenthesized list of possible values
to the `help` text, another was to present alternatives as part of the
metavariable. This commit standardizes on the latter, and generally
unifies more handling of enumerations in the CLI.