Bump asl-translator, macaw submodules to fix pop/ldm semantics
This bumps submodules to bring in the following changes:

* https://github.com/GaloisInc/asl-translator/pull/58
* https://github.com/GaloisInc/macaw/pull/547

These fix the `macaw-aarch32` semantics for instructions like `pop` and `ldm`
to avoid performing unnecessary memory reads, thereby preventing the issue that
was causing the `thumb-calls` test case to incorrectly be flagged as performing
an uninitialized stack read. As such, this fixes #432.

Note that `thumb-calls` still doesn't pass `grease` even with these changes,
but this is because `macaw`'s code discovery is not smart enough to reckon with
a function that ends with a `pop {pc, ...}` statement. Fixing that problem on
the `macaw` side would be non-trivial, so I have opted to accept this.