Add test case for macaw-aarch32 pop semantics
`grease` allows us to robustly check if the `macaw-aarch32` semantics for the
`pop` instruction performs spurious reads from uninitialized parts of the stack
or not. To ensure that the test case always includes a `pop` instruction of
just the right shape, we use an assembly file as the source.

Fixes https://github.com/GaloisInc/macaw/issues/543.